% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{discourse-package}
\alias{discourse}
\alias{discourse-package}
\title{discourse: Data-simulation via Iterative Stochastic Combinatorial Optimization Using Reported Summary Estimates}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

I introduce DISCOURSE – this algorithmic framework is designed to reconstruct complete data sets solely from summary statistics. It provides researchers with a tool to follow-up analyses even when raw data are unavailable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sebastian-lortz/discourse}
  \item \url{https://sebastian-lortz.github.io/discourse/}
  \item Report bugs at \url{https://github.com/sebastian-lortz/discourse/issues}
}

}
\author{
\strong{Maintainer}: Sebastian A. J. Lortz \email{s.a.j.lortz@gmail.com} (\href{https://orcid.org/0009-0007-0281-7996}{ORCID})

}
