% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{nds3-package}
\alias{nds3}
\alias{nds3-package}
\title{nds3: Non-Parametric Data Simulation from Summary Statistics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

We introduce the NDS3 framework: Nonparametric Data Simulation from Summary Statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sebastian-lortz/nds3}
  \item \url{https://sebastian-lortz.github.io/nds3/}
  \item Report bugs at \url{https://github.com/sebastian-lortz/nds3/issues}
}

}
\author{
\strong{Maintainer}: Sebastian A. J. Lortz \email{s.a.j.lortz@gmail.com} (\href{https://orcid.org/0009-0007-0281-7996}{ORCID})

}
