% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_cooling.R
\name{plot_cooling}
\alias{plot_cooling}
\title{Plot cooling schedule of a nds3.object}
\usage{
plot_cooling(nds3_obj)
}
\arguments{
\item{nds3_obj}{A `nds3.object` returned by one of the `optim_*` functions, containing
`inputs$max_iter`, `inputs$init_temp`, and `inputs$cooling_rate`.}
}
\value{
A `ggplot2` object.
}
\description{
Visualizes the squared annealing temperature across iterations for a given
`nds3.object`, allowing inspection of the cooling schedule used in optimization.
}
\examples{
 \dontrun{
result <- optim_aov(args = ..., ...)
plot_cooling(result)
}
}
