% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_rmse.R
\name{plot_rmse}
\alias{plot_rmse}
\title{Plot RMSE Comparison for discourse.object runs}
\usage{
plot_rmse(object_list)
}
\arguments{
\item{object_list}{A `discourse.object` or list thereof, typically output from `optim_*` functions.}
}
\value{
A `ggplot2` object.
}
\description{
Visualizes the root-mean-square error (RMSE) distributions for between-run variability
versus deviations from target values across multiple `discourse.object` runs (e.g. parallel_aov()).
}
\examples{
 \dontrun{
results <- parallel_aov(args = ..., ...)
plot_rmse(results)
}
}
