% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_print.R
\name{print.summary.nds3.object}
\alias{print.summary.nds3.object}
\title{Print summary for summary.nds3.object}
\usage{
\method{print}{summary.nds3.object}(x, ...)
}
\arguments{
\item{x}{A `summary.nds3.object` produced by `summary.nds3.object`.}

\item{...}{Unused. Additional arguments are ignored.}
}
\value{
Invisibly returns the input object `x` after printing its summary.
}
\description{
A minimal S3 print method for objects of class `summary.nds3.object`, providing
optimization diagnostics and summary statistics in the console.
}
