% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_summary.R
\name{summary.discourse.object}
\alias{summary.discourse.object}
\title{Summarize a discourse.object}
\usage{
\method{summary}{discourse.object}(object, ...)
}
\arguments{
\item{object}{A `discourse.object` produced by `optim_*` functions.}

\item{...}{Additional arguments (unused).}
}
\value{
An object of class `summary.discourse.object` containing:
\describe{
  \item{rmse}{Named list of RMSE values.}
  \item{inputs}{Original input list.}
  \item{data}{Optimized data frame.}
  \item{best_error}{Numeric. Minimum objective error achieved.}
  \item{track_error}{Numeric vector of best error at each iteration.}
  \item{track_error_ratio}{Numeric vector of error ratios (cor vs. reg) per iteration (if available).}
  \item{model}{Fitted regression model or ANOVA table (if applicable).}
  \item{coefficients}{regression estimates (LM/LME only).}
  \item{std_errors}{Coefficient standard errors (LM/LME only).}
  \item{correlations}{Correlation values (LM/LME only).}
  \item{F_value}{ANOVA F statistics (ANOVA only).}
  \item{means}{Group or variable means.}
  \item{sds}{Group or variable standard deviations (vector only).}
}
}
\description{
Provides a comprehensive summary of a `discourse.object`, including RMSE,
model components, and statistics depending on the analysis type (LM/LME, ANOVA, or vec).
}
\examples{
 \dontrun{
res_lm <- optim_lm(args = ..., ...)
summary(res_lm)
}
}
