% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_summary.R
\name{summary.nds3.object}
\alias{summary.nds3.object}
\title{Summarize a nds3.object}
\usage{
\method{summary}{nds3.object}(object, ...)
}
\arguments{
\item{object}{A `nds3.object` produced by `optim_*` functions.}

\item{...}{Additional arguments (unused).}
}
\value{
An object of class `summary.nds3.object` containing:
\describe{
  \item{rmse}{Named list of RMSE values.}
  \item{inputs}{Original input list.}
  \item{data}{Optimized data frame.}
  \item{best_error}{Numeric. Minimum objective error achieved.}
  \item{track_error}{Numeric vector of best error at each iteration.}
  \item{track_error_ratio}{Numeric vector of error ratios (cor vs. reg) per iteration (if available).}
  \item{model}{Fitted regression model or ANOVA table (if applicable).}
  \item{coefficients}{regression estimates (LM only).}
  \item{std_errors}{Coefficient standard errors (LM only).}
  \item{correlations}{Correlation values (LM only).}
  \item{F_value}{ANOVA F statistics (ANOVA only).}
  \item{means}{Group or variable means.}
  \item{sds}{Group or variable standard deviations (vector only).}
}
}
\description{
Provides a comprehensive summary of a `nds3.object`, including RMSE,
model components, and statistics depending on the analysis type (LM, ANOVA, or vec).
}
\examples{
 \dontrun{
res_lm <- optim_lm(args = ..., ...)
summary(res_lm)
}
}
